+++ examples/vpc_resource/vpc.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     oxide = {
       source  = "oxidecomputer/oxide"
-      version = "0.6.0"
+      version = "0.7.0"
     }
   }
 }